+++ tools/benchmark_test.py
@@ -4,7 +4,7 @@ import sys
 import os
 import benchmark
 import unittest
-from util import DenoTestCase, test_main
+from test_util import DenoTestCase, run_tests
 
 
 class TestBenchmark(DenoTestCase):
@@ -56,4 +56,4 @@ class TestBenchmark(DenoTestCase):
 if __name__ == '__main__':
     # FIME this doesn't appear to be the case.
     # This test assumes tools/http_server.py is running in the background.
-    test_main()
+    run_tests()
