+++ tools/repl_test.py
@@ -4,18 +4,10 @@ from subprocess import CalledProcessError, PIPE, Popen
 import sys
 import time
 
-from util import DenoTestCase, test_main
+from test_util import DenoTestCase, run_tests
 
 
 class TestRepl(DenoTestCase):
-    def __init__(self, *args, **kwargs):
-        super(TestRepl, self).__init__(*args, **kwargs)
-        self._warm_up()
-
-    def _warm_up(self):
-        # This may output an error message about the history file (ignore it).
-        self.input("")
-
     def input(self, *lines, **kwargs):
         exit_ = kwargs.pop("exit", True)
         sleep_ = kwargs.pop("sleep", 0)
@@ -141,4 +133,4 @@ class TestRepl(DenoTestCase):
 
 
 if __name__ == "__main__":
-    test_main()
+    run_tests()
